+++ app/models.py
@@ -1,8 +1,11 @@
-from app import db
 from flask_login import UserMixin
 from werkzeug.security import check_password_hash, generate_password_hash
 from datetime import datetime
 import hashlib
+from flask_sqlalchemy import SQLAlchemy
+
+
+db = SQLAlchemy()
 
 
 class User(UserMixin, db.Model):
