--- app/reports/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-bp = Blueprint('reports', __name__)
-
-from app.reports import routes
